requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Deep Learning & Computer Vision
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-image==0.22.0

# Face Recognition Models
insightface==0.7.3
facenet-pytorch==2.5.3

# Search & Indexing
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Data Processing
pandas==2.1.3
scikit-learn==1.3.2
albumentations==1.3.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring & Metrics
prometheus-client==0.19.0
